requirements.txt
# ZenAi Dependencies / ZenAi 依赖

# Core LLM integration / 核心 LLM 集成
openai>=2.0.0

# Environment variables / 环境变量
python-dotenv>=1.0.0

# Configuration / 配置
pyyaml>=6.0.0

# Database / 数据库
sqlalchemy>=2.0.0

# API Framework / API 框架
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0

# Scheduling / 调度
apscheduler>=3.10.0

# Testing / 测试
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-mock>=3.12.0
httpx>=0.26.0  # For testing FastAPI
aiohttp>=3.9.0  # For async HTTP requests
